feat:download plugin

+++ src-tauri/src/plugin.rs
@@ -3,20 +3,12 @@ use std::{
     path::Path,
 };
 
+use crate::entity::plugin::{ActiveModel, Entity, StringVec};
 use crate::utils::command;
 use crate::{error::MyError, git::Git};
 use git2::Progress;
+use sea_orm::*;
 use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub enum InstallType {
-    #[serde(rename = "git-clone")]
-    GitClone,
-    #[serde(rename = "unzip")]
-    UnZip,
-    #[serde(rename = "copy")]
-    Copy,
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Plugin {
@@ -25,8 +17,29 @@ pub struct Plugin {
     pub reference: String,
     pub pip: Option<Vec<String>>,
     pub files: Vec<String>,
-    pub install_type: InstallType,
+    pub install_type: String,
     pub description: String,
+}
+
+impl IntoActiveModel<ActiveModel> for Plugin {
+    fn into_active_model(self) -> ActiveModel {
+        ActiveModel {
+            title: Set(self.title),
+            author: Set(self.author),
+            reference: Set(self.reference),
+            description: Set(self.description),
+            install_type: Set(self.install_type),
+            pip: Set(self.pip.map(|p| p.into())),
+            files: Set(self.files.into()),
+            ..Default::default()
+        }
+    }
+}
+
+impl From<Vec<String>> for StringVec {
+    fn from(value: Vec<String>) -> Self {
+        Self(value)
+    }
 }
 
 impl Plugin {
@@ -85,12 +98,10 @@ impl Plugin {
 
     pub async fn remove(&self, comfyui_path: &str) -> Result<(), MyError> {
         let name = self.get_file_name();
-        let path = Path::new(comfyui_path)
-            .join("custom_nodes")
-            .join(name)
-            .display()
-            .to_string();
-        std::fs::remove_dir_all(path)?;
+        let path = Path::new(comfyui_path).join("custom_nodes").join(name);
+        if path.exists() {
+            std::fs::remove_dir_all(path)?;
+        }
         Ok(())
     }
 }
@@ -120,4 +131,15 @@ impl PluginList {
         let plugin_list: PluginList = serde_json::from_reader(file)?;
         Ok(plugin_list)
     }
+}
+
+pub async fn init_plugin_list(db: &DbConn, plugin_list: PluginList) -> anyhow::Result<()> {
+    Entity::delete_many().exec(db).await?;
+    let models = plugin_list
+        .custom_nodes
+        .into_iter()
+        .map(|p| p.into_active_model())
+        .collect::<Vec<_>>();
+    Entity::insert_many(models).exec(db).await?;
+    Ok(())
 }
